clean cp tests
