+++ crates/nu-command/tests/commands/mod.rs
@@ -12,7 +12,6 @@ mod config_env_default;
 mod config_nu_default;
 mod continue_;
 mod conversions;
-mod cp;
 #[cfg(feature = "sqlite")]
 mod database;
 mod date;
